Move changes to the correct files
